Development

+++ lib/src/views/screens/library/uploads/uploaded_songs.dart
@@ -1,19 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:get/get.dart';
 import 'package:nuphonic_front_end/src/app_logics/models/song_model.dart';
 import 'package:nuphonic_front_end/src/app_logics/services/api_services/song_service.dart';
 import 'package:nuphonic_front_end/src/app_logics/services/shared_pref_services/shared_pref_service.dart';
 import 'package:nuphonic_front_end/src/views/reusable_widgets/custom_error.dart';
+import 'package:nuphonic_front_end/src/views/reusable_widgets/custom_refresh_header.dart';
 import 'package:nuphonic_front_end/src/views/reusable_widgets/custom_snackbar.dart';
 import 'package:nuphonic_front_end/src/views/reusable_widgets/song_box.dart';
+import 'package:nuphonic_front_end/src/views/screens/library/upload_song.dart';
 import 'package:nuphonic_front_end/src/views/utils/consts.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
 
 class UploadedSongs extends StatefulWidget {
   @override
   _UploadedSongsState createState() => _UploadedSongsState();
 }
 
-class _UploadedSongsState extends State<UploadedSongs> {
+class _UploadedSongsState extends State<UploadedSongs>
+    with AutomaticKeepAliveClientMixin<UploadedSongs> {
+  @override
+  bool get wantKeepAlive => true;
+
+  RefreshController _refreshController = RefreshController();
   SongService _songService = SongService();
   SharedPrefService _sharedPrefService = SharedPrefService();
   CustomSnackBar _customSnackBar = CustomSnackBar();
@@ -45,31 +54,38 @@ class _UploadedSongsState extends State<UploadedSongs> {
   }
 
   Future<void> deleteSong(SongModel song) async {
-    setState(() {
-      isLoading = true;
-    });
-    var userID = await _sharedPrefService.read(id: 'user_id');
-    dynamic result =
-    await _songService.deleteSong(userID, song.songID);
-    setState(() {
-      isLoading = false;
-    });
-    if (result == null) {
-      _customSnackBar.buildSnackBar('Network Error', false);
-    } else {
-      _customSnackBar.buildSnackBar(result.data['msg'], result.data['success']);
-      if (result.data['success']) {
-        getUserSongs();
+    if (song.albumName == 'Single') {
+      setState(() {
+        isLoading = true;
+      });
+      var userID = await _sharedPrefService.read(id: 'user_id');
+      dynamic result = await _songService.deleteSong(userID, song.songID);
+      setState(() {
+        isLoading = false;
+      });
+      if (result == null) {
+        _customSnackBar.buildSnackBar('Network Error', false);
+      } else {
+        _customSnackBar.buildSnackBar(
+            result.data['msg'], result.data['success']);
+        if (result.data['success']) {
+          getUserSongs();
+        }
       }
+    } else {
+      _customSnackBar.buildSnackBar(
+          'Remove this song from album first in order to delete', false);
     }
   }
 
   Widget _showErrorMessage() {
     return CustomError(
       title: 'No Songs Found',
-      subTitle:
-          'You have not uploaded any song yet. Upload song to see here.',
+      subTitle: 'You have not uploaded any song yet. Upload song to see here.',
       buttonLabel: 'UPLOAD',
+      onPressed: () {
+        Get.to(UploadSongs());
+      },
     );
   }
 
@@ -112,16 +128,26 @@ class _UploadedSongsState extends State<UploadedSongs> {
 
   @override
   Widget build(BuildContext context) {
-    return isLoading ? loading : _uploadedSongs.length == 0
-        ? _showErrorMessage()
-        : SingleChildScrollView(
-            child: Column(children: [
-              SizedBox(
-                height: 20,
-              ),
-              _showSongs()
-            ]),
-          );
+    return isLoading
+        ? loading
+        : _uploadedSongs.length == 0
+            ? _showErrorMessage()
+            : SmartRefresher(
+                controller: _refreshController,
+                onRefresh: () {
+                  getUserSongs()
+                      .then((value) => _refreshController.refreshCompleted());
+                },
+                header: CustomRefreshHeader(),
+                child: SingleChildScrollView(
+                  child: Column(children: [
+                    SizedBox(
+                      height: 20,
+                    ),
+                    _showSongs()
+                  ]),
+                ),
+              );
     ;
   }
 }
